Incorporating feedback
