+++ Cargo.toml
@@ -34,6 +34,7 @@ self_update = "0.5.0"
 num-derive = "0.2.3"
 num-traits = "0.2.6"
 annotate-snippets = { version = "0.5.0", features = ["ansi_term"] }
+atty = "0.2.0"
 
 [dev-dependencies]
 assert_cli = "0.6.3"
