+++ src/lang.rs
@@ -1,11 +1,10 @@
 use crate::data;
-use annotate_snippets::snippet::{Annotation, AnnotationType, Slice, Snippet, SourceAnnotation};
+use crate::errors::SyntaxErrors;
 use nom;
 use nom::types::CompleteStr;
 use nom::*;
 use nom::{digit1, double, is_alphabetic, is_alphanumeric, is_digit, multispace};
 use nom_locate::LocatedSpan;
-use num_traits::FromPrimitive;
 use std::convert::From;
 use std::str;
 
@@ -53,192 +52,6 @@ pub struct Positioned<T> {
 impl<T> Positioned<T> {
     pub fn into(&self) -> &T {
         &self.value
-    }
-}
-
-/// Common syntax errors.
-#[derive(PartialEq, Debug, FromPrimitive, Fail)]
-pub enum SyntaxErrors {
-    #[fail(display = "")]
-    DelimiterStart,
-    #[fail(display = "unterminated double quote string")]
-    UnterminatedString,
-    #[fail(display = "expecting close parentheses")]
-    MissingParen,
-}
-
-impl SyntaxErrors {
-    fn to_resolution(&self) -> &'static str {
-        match self {
-            SyntaxErrors::DelimiterStart => "",
-            SyntaxErrors::UnterminatedString => "Insert a double quote to terminate this string",
-            SyntaxErrors::MissingParen => "Insert a right parenthesis to close this expression",
-        }
-    }
-}
-
-/// Converts the ordinal from the nom error object back into a SyntaxError.
-impl From<u32> for SyntaxErrors {
-    fn from(ord: u32) -> Self {
-        SyntaxErrors::from_u32(ord).unwrap()
-    }
-}
-
-impl From<SyntaxErrors> for ErrorKind {
-    fn from(e: SyntaxErrors) -> Self {
-        ErrorKind::Custom(e as u32)
-    }
-}
-
-/// Callback for handling error Snippets.
-pub trait ErrorReporter {
-    fn handle_error(&self, _snippet: Snippet) {}
-}
-
-/// Container for the query string that can be used to parse and report errors.
-pub struct QueryContainer {
-    query: String,
-    reporter: Box<ErrorReporter>,
-}
-
-impl QueryContainer {
-    pub fn new(query: String, reporter: Box<ErrorReporter>) -> QueryContainer {
-        QueryContainer { query, reporter }
-    }
-
-    /// Create a SnippetBuilder for the given error
-    pub fn report_error_for<E: ToString>(&self, error: E) -> SnippetBuilder {
-        SnippetBuilder {
-            query: self,
-            data: SnippetData {
-                error: error.to_string(),
-                source: self.query.to_string(),
-                ..Default::default()
-            },
-        }
-    }
-
-    /// Parse the contained query string.
-    pub fn parse(&self) -> Result<Query, QueryPosition> {
-        let parse_result = query(Span::new(CompleteStr(&self.query)));
-
-        match parse_result {
-            Err(nom::Err::Failure(nom::Context::List(ref list))) => {
-                // Check for an error from a delimited!() parser.  The error list will contain
-                // the location of the start as the last item and the location of the end as the
-                // penultimate item.
-                let last_chunk = list.rchunks_exact(2).next().map(|p| (&p[0], &p[1]));
-
-                match last_chunk {
-                    Some((
-                        (ref end_span, ErrorKind::Custom(ref delim_error)),
-                        (ref start_span, ErrorKind::Custom(SyntaxErrors::DelimiterStart)),
-                    )) => {
-                        self.report_error_for(delim_error)
-                            .with_annotation_range((*start_span).into(), (*end_span).into(), "")
-                            .with_resolution(delim_error.to_resolution())
-                            .send_report();
-                    }
-                    _ => self.report_error_for(format!("{:?}", list)).send_report(),
-                }
-            }
-            Err(nom::Err::Error(nom::Context::Code(span, _))) => {
-                self.report_error_for("Unexpected input")
-                    .with_annotation_range(span.into(), QueryPosition(self.query.len()), "")
-                    .send_report();
-            }
-            _ => (),
-        }
-        // Return the parsed value or the last position of valid syntax
-        parse_result.map(|x| x.1).map_err(|e| match e {
-            nom::Err::Incomplete(_) => QueryPosition(0),
-            nom::Err::Error(context) | nom::Err::Failure(context) => match context {
-                nom::Context::Code(span, _) => span.into(),
-                nom::Context::List(list) => list.first().unwrap().0.into(),
-            },
-        })
-    }
-}
-
-/// Container for data that will be used to construct a Snippet
-#[derive(Default)]
-pub struct SnippetData {
-    error: String,
-    source: String,
-    annotations: Vec<((usize, usize), String)>,
-    resolution: Vec<String>,
-}
-
-pub struct SnippetBuilder<'a> {
-    query: &'a QueryContainer,
-    data: SnippetData,
-}
-
-impl<'a> SnippetBuilder<'a> {
-    /// Adds an annotation to a portion of the query string.  The given position will be
-    /// highlighted with the accompanying label.
-    pub fn with_annotation<T, S: ToString>(mut self, pos: &Positioned<T>, label: S) -> Self {
-        self.data
-            .annotations
-            .push(((pos.start_pos.0, pos.end_pos.0), label.to_string()));
-        self
-    }
-
-    /// Adds an annotation to a portion of the query string.  The given position will be
-    /// highlighted with the accompanying label.
-    pub fn with_annotation_range<S: ToString>(
-        mut self,
-        start_pos: QueryPosition,
-        end_pos: QueryPosition,
-        label: S,
-    ) -> Self {
-        self.data
-            .annotations
-            .push(((start_pos.0, end_pos.0), label.to_string()));
-        self
-    }
-
-    /// Add a message to help the user resolve the error.
-    pub fn with_resolution<T: ToString>(mut self, resolution: T) -> Self {
-        self.data.resolution.push(resolution.to_string());
-        self
-    }
-
-    /// Build and send the Snippet to the ErrorReporter in the QueryContainer.
-    pub fn send_report(mut self) {
-        self.query.reporter.handle_error(Snippet {
-            title: Some(Annotation {
-                label: Some(self.data.error),
-                id: None,
-                annotation_type: AnnotationType::Error,
-            }),
-            slices: vec![Slice {
-                source: self.data.source,
-                line_start: 1,
-                origin: None,
-                fold: false,
-                annotations: self
-                    .data
-                    .annotations
-                    .drain(..)
-                    .map(move |anno| SourceAnnotation {
-                        range: anno.0,
-                        label: anno.1,
-                        annotation_type: AnnotationType::Error,
-                    })
-                    .collect(),
-            }],
-            footer: self
-                .data
-                .resolution
-                .iter()
-                .map(|res| Annotation {
-                    label: Some(res.to_string()),
-                    id: None,
-                    annotation_type: AnnotationType::Help,
-                })
-                .collect(),
-        });
     }
 }
 
@@ -694,7 +507,7 @@ named!(filter<Span, Vec<Keyword>>, map!(
     }
 ));
 
-named!(query<Span, Query, SyntaxErrors>, fix_error!(SyntaxErrors, exact!(ws!(do_parse!(
+named!(pub query<Span, Query, SyntaxErrors>, fix_error!(SyntaxErrors, exact!(ws!(do_parse!(
     filter: filter >>
     operators: opt!(preceded!(tag!("|"), ws!(separated_nonempty_list!(tag!("|"), operator)))) >>
     (Query{
